Added a proper main activity which displays that it is loading.
- If it finds a user is logged in, it will direct them to their appropriate activity. Otherwise, it will go to the login activity.
