+++ app/src/main/AndroidManifest.xml
@@ -1,35 +1,44 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.arom.jobzi">
+          package="com.arom.jobzi">
 
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
+	<uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
+	<uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
 
-    <application
-        android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
-        android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-        <activity android:name=".LoginActivity" android:noHistory="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+	<application
+		android:allowBackup="true"
+		android:icon="@mipmap/ic_launcher"
+		android:label="@string/app_name"
+		android:roundIcon="@mipmap/ic_launcher_round"
+		android:supportsRtl="true"
+		android:theme="@style/AppTheme">
+		<activity
+			android:name=".MainActivity"
+			android:noHistory="true">
+			<intent-filter>
+				<action android:name="android.intent.action.MAIN"/>
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
-            android:name=".SignupActivity"
-            android:label="@string/title_activity_signup"
-            android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".ServiceEditorActivity" />
-        <activity
-            android:name=".AdminActivity"
-            android:label="@string/title_activity_admin"
-            android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".HomeOwnerActivity" />
-        <activity android:name=".ServiceProviderActivity" />
-    </application>
+				<category android:name="android.intent.category.LAUNCHER"/>
+			</intent-filter>
+		</activity>
+		<activity
+			android:name=".LoginActivity"
+			android:label="@string/title_activity_login">
+		</activity>
+		<activity
+			android:name=".SignupActivity"
+			android:label="@string/title_activity_signup"
+			android:theme="@style/AppTheme.NoActionBar"/>
+		<activity
+			android:name=".ServiceEditorActivity"
+			android:label="@string/title_activity_service_editor"/>
+		<activity
+			android:name=".AdminActivity"
+			android:label="@string/title_activity_admin"
+			android:theme="@style/AppTheme.NoActionBar"/>
+		<activity android:name=".HomeOwnerActivity"/>
+		<activity android:name=".ServiceProviderActivity"/>
+
+	</application>
 
 </manifest>
