+++ app/src/main/java/com/arom/jobzi/MainActivity.java
@@ -1,0 +1,28 @@
+package com.arom.jobzi;
+
+import android.content.Intent;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.arom.jobzi.util.Util;
+import com.google.firebase.auth.FirebaseUser;
+
+public class MainActivity extends AppCompatActivity {
+	
+	@Override
+	protected void onCreate(Bundle savedInstanceState) {
+		super.onCreate(savedInstanceState);
+		setContentView(R.layout.activity_main);
+		
+		FirebaseUser currentUser = Util.getInstance().getCurrentUser();
+		
+		if(currentUser != null) {
+			Util.getInstance().onUserLogin(this, currentUser);
+		} else {
+			Intent toLoginIntent = new Intent(this, LoginActivity.class);
+			startActivity(toLoginIntent);
+			finish();
+		}
+		
+	}
+}
