+++ app/src/main/res/layout/activity_login.xml
@@ -32,8 +32,8 @@
 
                 <ImageView
                     android:id="@+id/imageView1"
-                    android:layout_width="122dp"
-                    android:layout_height="96dp"
+                    android:layout_width="@dimen/logo_width"
+                    android:layout_height="@dimen/logo_height"
                     android:layout_marginStart="8dp"
                     android:layout_marginTop="8dp"
                     android:layout_marginEnd="8dp"
