+++ app/src/main/res/layout/activity_main.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout
+	xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	xmlns:tools="http://schemas.android.com/tools"
+	android:layout_width="match_parent"
+	android:layout_height="match_parent"
+	tools:context=".MainActivity">
+
+	<ImageView
+		android:id="@+id/logoImageView"
+		android:layout_width="@dimen/logo_width"
+		android:layout_height="@dimen/logo_height"
+		android:layout_centerHorizontal="true"
+		android:layout_centerVertical="true"
+		android:src="@drawable/proxy_brush"
+		/>
+
+	<ProgressBar
+		android:id="@+id/progressBar"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:layout_below="@id/logoImageView"
+		android:layout_marginTop="16dp"
+		android:layout_centerHorizontal="true"
+		/>
+</RelativeLayout>
