+++ app/src/main/res/values/dimens.xml
@@ -6,4 +6,6 @@
     <dimen name="activity_vertical_margin">16dp</dimen>
     <dimen name="appbar_padding_top">8dp</dimen>
     <dimen name="text_margin">16dp</dimen>
+    <dimen name="logo_width">122dp</dimen>
+    <dimen name="logo_height">96dp</dimen>
 </resources>
