+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,9 @@
 <resources>
 	<string name="app_name">JobZi</string>
-	<string name="title_activity_welcome">WelcomeActivity</string>
+	<string name="title_activity_login">LoginActivity</string>
 	<string name="title_activity_signup">SignupActivity</string>
 	<string name="title_activity_admin">AdminActivity</string>
+	<string name="title_activity_service_editor">ServiceEditorActivity</string>
 	<string name="email">Email</string>
 	<string name="username">Username</string>
 	<string name="password">Password</string>
@@ -20,8 +21,8 @@
 	<string name="cancel">Cancel</string>
 	<string name="rate_hint_text">Rate ($/hour)</string>
 	<string name="service_name_hint_text">Service Name</string>
-    <string name="navigation_drawer_open">Open navigation drawer</string>
-    <string name="navigation_drawer_close">Close navigation drawer</string>
+	<string name="navigation_drawer_open">Open navigation drawer</string>
+	<string name="navigation_drawer_close">Close navigation drawer</string>
 	<string name="user_preferences">User Preferences</string>
 
 </resources>
